+++ application/controllers/Mtce.php
@@ -19,16 +19,21 @@ class Mtce extends Application {
     {
         $this->data['pagetitle'] = 'TODO List Maintenance';
         // build the task presentation output
-        $result = ''; // start with an empty array      
+        $result = ''; // start with an empty array    
+        
+        $role = $this->session->userdata('userrole');
         foreach ($tasks as $task)
         {
             if (!empty($task->status))
                 $task->status = $this->app->status($task->status);
-            $result .= $this->parser->parse('oneitem', (array) $task, true);
+            
+            if ($role == ROLE_OWNER)
+                $result .= $this->parser->parse('oneitemx', (array) $task, true);
+            else
+                $result .= $this->parser->parse('oneitem', (array) $task, true);
         }
         $this->data['display_tasks'] = $result;
 
-        $role = $this->session->userdata('userrole');
         $this->data['pagetitle'] = 'TODO List Maintenance ('. $role . ')';
 
         // and then pass them on
@@ -54,8 +59,11 @@ class Mtce extends Application {
             if ($count >= $this->items_per_page) break;
         }  
         $this->data['pagination'] = $this->pagenav($num);
-        $this->show_page($tasks);
         
+        $role = $this->session->userdata('userrole');
+        if ($role == ROLE_OWNER) 
+            $this->data['pagination'] .= $this->parser->parse('itemadd',[], true);
+        $this->show_page($tasks);        
     }
     
     // Build the pagination navbar