+++ application/views/oneitemx.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: Getry
+ * Date: 2017-10-19
+ * Time: 2:20 PM
+ */
+?>
+
+<tr>
+    <td><a href="/mtce/edit/{id}"><input type="button" value="{id}"/></a></td>
+    <td>{task}</td>
+    <td>{status}</td>
+</tr>